[+] 完善结构，抽离行为树

+++ Assets/Plugins/GamesTan/Lockstep.AI.BehaviourTree/Runtime/Node/BTActionParallel.cs
@@ -106,7 +106,7 @@ namespace Lockstep.AI {
             }
 
             //clear running status
-            thisContext->StatusRunning.Init((byte) (int) BTRunningStatus.EXECUTING);
+            thisContext->StatusRunning.Init( BTRunningStatus.EXECUTING);
         }
     }
 }
